requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Werkzeug==3.0.6
mysql-connector-python==9.1.0
python-dotenv
gunicorn
